Add rollup_finding flag for rolled-up findings
