+++ src/models.rs
@@ -7,6 +7,7 @@
 pub mod attachment;
 pub mod family_selection;
 pub mod host_property;
+pub mod item;
 pub mod plugin_metadata;
 pub mod plugin_preference;
 pub mod policy;
@@ -18,6 +19,7 @@ pub mod service_description;
 pub use attachment::Attachment;
 pub use family_selection::FamilySelection;
 pub use host_property::HostProperty;
+pub use item::Item;
 pub use plugin_metadata::NessusPluginMetadata;
 pub use plugin_preference::PluginPreference;
 pub use policy::Policy;
@@ -30,7 +32,7 @@ use diesel::prelude::*;
 use diesel::sqlite::SqliteConnection;
 use std::net::IpAddr;
 
-use crate::schema::{nessus_hosts, nessus_items, nessus_patches, nessus_plugins};
+use crate::schema::{nessus_hosts, nessus_patches, nessus_plugins};
 
 #[derive(Debug, Queryable, Identifiable)]
 #[diesel(table_name = nessus_hosts)]
@@ -49,80 +51,6 @@ pub struct Host {
     pub risk_score: Option<i32>,
     pub user_id: Option<i32>,
     pub engagement_id: Option<i32>,
-}
-
-#[derive(Debug, Queryable, Identifiable, Associations)]
-#[diesel(belongs_to(Host, foreign_key = host_id))]
-#[diesel(belongs_to(Plugin, foreign_key = plugin_id))]
-#[diesel(table_name = nessus_items)]
-pub struct Item {
-    pub id: i32,
-    pub host_id: Option<i32>,
-    pub plugin_id: Option<i32>,
-    pub attachment_id: Option<i32>,
-    pub plugin_output: Option<String>,
-    pub port: Option<i32>,
-    pub svc_name: Option<String>,
-    pub protocol: Option<String>,
-    pub severity: Option<i32>,
-    pub plugin_name: Option<String>,
-    pub description: Option<String>,
-    pub solution: Option<String>,
-    pub risk_factor: Option<String>,
-    pub cvss_base_score: Option<f32>,
-    pub verified: Option<bool>,
-    pub cm_compliance_info: Option<String>,
-    pub cm_compliance_actual_value: Option<String>,
-    pub cm_compliance_check_id: Option<String>,
-    pub cm_compliance_policy_value: Option<String>,
-    pub cm_compliance_audit_file: Option<String>,
-    pub cm_compliance_check_name: Option<String>,
-    pub cm_compliance_result: Option<String>,
-    pub cm_compliance_output: Option<String>,
-    pub cm_compliance_reference: Option<String>,
-    pub cm_compliance_see_also: Option<String>,
-    pub cm_compliance_solution: Option<String>,
-    pub real_severity: Option<i32>,
-    pub risk_score: Option<i32>,
-    pub user_id: Option<i32>,
-    pub engagement_id: Option<i32>,
-}
-
-impl Default for Item {
-    fn default() -> Self {
-        Self {
-            id: 0,
-            host_id: None,
-            plugin_id: None,
-            attachment_id: None,
-            plugin_output: None,
-            port: None,
-            svc_name: None,
-            protocol: None,
-            severity: None,
-            plugin_name: None,
-            description: None,
-            solution: None,
-            risk_factor: None,
-            cvss_base_score: None,
-            verified: None,
-            cm_compliance_info: None,
-            cm_compliance_actual_value: None,
-            cm_compliance_check_id: None,
-            cm_compliance_policy_value: None,
-            cm_compliance_audit_file: None,
-            cm_compliance_check_name: None,
-            cm_compliance_result: None,
-            cm_compliance_output: None,
-            cm_compliance_reference: None,
-            cm_compliance_see_also: None,
-            cm_compliance_solution: None,
-            real_severity: None,
-            risk_score: None,
-            user_id: None,
-            engagement_id: None,
-        }
-    }
 }
 
 #[derive(Debug, Queryable, Identifiable)]
